RestoreAnalogTitleScroll: slight refactor
moved the IDSystem stuff to a new game_ui.h, as virtually every hud element makes use of these
