+++ dllmain/dllmain.vcxproj
@@ -1029,6 +1029,7 @@ echo | set /p dummyName=#define GIT_BRANCH ^"%VAR%^" &gt;&gt; gitparams.h</Comma
     <ClInclude Include="SDK\event.h" />
     <ClInclude Include="SDK\fade.h" />
     <ClInclude Include="SDK\filter00.h" />
+    <ClInclude Include="SDK\game_ui.h" />
     <ClInclude Include="SDK\gc_math.h" />
     <ClInclude Include="SDK\global.h" />
     <ClInclude Include="SDK\GX.h" />
