+++ dllmain/dllmain.vcxproj.filters
@@ -1599,6 +1599,9 @@
     <ClInclude Include="SDK\fade.h">
       <Filter>dllmain\SDK</Filter>
     </ClInclude>
+    <ClInclude Include="SDK\game_ui.h">
+      <Filter>dllmain\SDK</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="includes">
